NNT Driver | PCI device discovery update
Description: to distinguish between PCI bar & PCI config devices we need to
check the pci device id range. this way we don't have to add each new device ad-hoc since the IDs
have a continuous pattern.
for the PPC driver we have to maintain a list of devices for the .id_table field of the driver